+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 
 ## 2.0 - unreleased
+- make relative URLs in e-* properties absolute (#201)
+- fix whitespace in plaintext conversion (#207)
+- add srcset support (#209)
+- add language support (#210)
 
 ## 1.1.3 - 2023-06-28
 - reduce instances where photo is implied (#135)